--- src/KokkosComm_collective.hpp
@@ -1,32 +1,0 @@
-//@HEADER
-// ************************************************************************
-//
-//                        Kokkos v. 4.0
-//       Copyright (2022) National Technology & Engineering
-//               Solutions of Sandia, LLC (NTESS).
-//
-// Under the terms of Contract DE-NA0003525 with NTESS,
-// the U.S. Government retains certain rights in this software.
-//
-// Part of Kokkos, under the Apache License v2.0 with LLVM Exceptions.
-// See https://kokkos.org/LICENSE for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//@HEADER
-
-#pragma once
-
-#include <Kokkos_Core.hpp>
-
-#include "KokkosComm_concepts.hpp"
-#include "KokkosComm_reduce.hpp"
-#include "KokkosComm_reducer.hpp"
-
-namespace KokkosComm {
-
-template <KokkosView SendView, KokkosView RecvView, KokkosExecutionSpace ExecSpace>
-void reduce(const ExecSpace &space, const SendView &sv, const RecvView &rv, Reducer op, int root, Communicator comm) {
-  return Impl::reduce(space, sv, rv, op, root, comm);
-}
-
-}  // namespace KokkosComm
